Split kubernetes flag type dependencies to sub-directories
