+++ flag/guest/cluster/kubernetes/hyperkube/docker/docker.go
@@ -1,0 +1,6 @@
+package docker
+
+// Docker is a data structure to hold Docker image configuration flag.
+type Docker struct {
+	Image string
+}
