+++ flag/guest/cluster/kubernetes/hyperkube/hyperkube.go
@@ -1,0 +1,9 @@
+package hyperkube
+
+import "github.com/giantswarm/kubernetesd/flag/service/cluster/kubernetes/hyperkube/docker"
+
+// Hyperkube is a data structure to hold guest cluster Kubernetes Hyperkube
+// image specific configuration flags.
+type Hyperkube struct {
+	Docker docker.Docker
+}
