+++ flag/guest/cluster/kubernetes/kubelet/kubelet.go
@@ -1,0 +1,9 @@
+package kubelet
+
+// Kubelet is a data structure to hold guest cluster kubelet specific
+// configuration flags.
+type Kubelet struct {
+	AltNames string
+	Labels   string
+	Port     string
+}
