+++ flag/guest/cluster/kubernetes/ssh/ssh.go
@@ -1,0 +1,7 @@
+package ssh
+
+// SSH is a data structure to hold guest cluster SSH specific configuration
+// flags.
+type SSH struct {
+	UserList string
+}
